requirements.txt
django
djangorestframework
psycopg2-binary
PyJWT
bcrypt
django-cors-headers
